# docs/swagger.yaml
definitions:
  models.AppDetails:
    properties:
      description:
        maxLength: 500
        minLength: 10
        type: string
      email:
        type: string
    required:
    - description
    - email
    type: object
...
  models.ApplicationRequest:
    properties:
      app_details:
        $ref: '#/definitions/models.AppDetails'
      app_name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - app_details
    - app_name
    type: object
  models.ApplicationResponse:
    properties:
      app_details:
        $ref: '#/definitions/models.AppDetails'
      app_did:
        type: string
      app_name:
        type: string
    type: object
...
  models.GetAccessTokenRequest:
    properties:
      app_did:
        type: string
      oauth_credential: {}
      policy_credential: {}
    required:
    - oauth_credential
    - policy_credential
    type: object
  models.GetAccessTokenResponse:
    properties:
      access_token:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /applications:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all applications
      parameters:
      - description: API Key
        in: header
        name: x-api-key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ApplicationResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all applications
      tags:
      - Application Management
    post:
      consumes:
      - application/json
      description: Create a new application with the given details
      parameters:
      - description: Application to create
        in: body
        name: application
        required: true
        schema:
          $ref: '#/definitions/models.ApplicationRequest'
      - description: API Key
        in: header
        name: x-api-key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApplicationResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new application
      tags:
      - Application Management
...
  /get-access-token:
    post:
      consumes:
      - application/json
      description: Handles the sign-in process using application DID, credential JWT,
        and session key.
      parameters:
      - default: Bearer YOUR_ACCESS_TOKEN
        description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Application to create
        in: body
        name: application
        required: true
        schema:
          $ref: '#/definitions/models.GetAccessTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Access Token
          schema:
            $ref: '#/definitions/models.GetAccessTokenResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Sign in or get access token to an application
      tags:
      - User Access Management
  /grant-access:
    put:
      consumes:
      - application/json
      description: Grants access based on a valid request.
      produces:
      - application/json
      responses:
        "200":
          description: implementation pending
          schema:
            type: string
        "405":
          description: Only PUT method is allowed
          schema:
            type: string
      summary: Grant access to a user
      tags:
      - Permission Management
...
  /request-access:
    post:
      consumes:
      - application/json
      description: Initiates a request for user access.
      produces:
      - application/json
      responses:
        "200":
          description: implementation pending
          schema:
            type: string
        "405":
          description: Only POST method is allowed
          schema:
            type: string
      summary: Request access for a user
      tags:
      - User Access Management
  /revoke-access:
    put:
      consumes:
      - application/json
      description: Revokes the access of a user.
      produces:
      - application/json
      responses:
        "200":
          description: implementation pending
          schema:
            type: string
        "405":
          description: Only PUT method is allowed
          schema:
            type: string
      summary: Revoke access of a user
      tags:
      - Permission Management
...
  /signup:
    get:
      consumes:
      - application/json
      description: Handles the sign-up process by providing a redirect URL for authentication.
      parameters:
      - description: Application DID
        in: query
        name: app_did
        required: true
        type: string
      - description: Session Token
        in: query
        name: session_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Redirect URL for sign-up
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Sign up for an application
      tags:
      - User Access Management
  /verify-access:
    get:
      consumes:
      - application/json
      description: Verifies if a user has access to a specific resource based on their
        role.
      parameters:
      - default: Bearer YOUR_ACCESS_TOKEN
        description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: e.g.; Role to check access for
        in: query
        name: attribute
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Verify access to a resource
      tags:
      - User Access Management
swagger: "2.0"